refactor: push more of the rpc logic down into the RpcServerManager / RpcClientManager layers
