+++ Sources/LiveKit/Core/RPC.swift
@@ -155,62 +155,3 @@ struct PendingRpcResponse {
     let participantIdentity: Participant.Identity
     let onResolve: @Sendable (_ payload: String?, _ error: RpcError?) -> Void
 }
-
-/// Caller-side RPC state: tracks pending acks and pending responses for outgoing requests.
-actor RpcClientManager: Loggable {
-    private var pendingAcks: Set<String> = Set()
-    private var pendingResponses: [String: PendingRpcResponse] = [:] // requestId to pending response
-
-    func addPendingAck(_ requestId: String) {
-        pendingAcks.insert(requestId)
-    }
-
-    @discardableResult
-    func removePendingAck(_ requestId: String) -> Bool {
-        pendingAcks.remove(requestId) != nil
-    }
-
-    func hasPendingAck(_ requestId: String) -> Bool {
-        pendingAcks.contains(requestId)
-    }
-
-    func setPendingResponse(_ requestId: String, response: PendingRpcResponse) {
-        pendingResponses[requestId] = response
-    }
-
-    @discardableResult
-    func removePendingResponse(_ requestId: String) -> PendingRpcResponse? {
-        pendingResponses.removeValue(forKey: requestId)
-    }
-
-    func removeAllPending(_ requestId: String) {
-        pendingAcks.remove(requestId)
-        pendingResponses.removeValue(forKey: requestId)
-    }
-}
-
-/// Handler-side RPC state: tracks registered method handlers for incoming requests.
-actor RpcServerManager: Loggable {
-    private var handlers: [String: RpcHandler] = [:] // methodName to handler
-
-    func registerHandler(_ method: String, handler: @escaping RpcHandler) throws {
-        guard !isRpcMethodRegistered(method) else {
-            throw LiveKitError(.invalidState, message: "RPC method '\(method)' already registered")
-        }
-        handlers[method] = handler
-    }
-
-    func unregisterHandler(_ method: String) {
-        if handlers.removeValue(forKey: method) == nil {
-            log("No handler registered for RPC method '\(method)'", .warning)
-        }
-    }
-
-    func isRpcMethodRegistered(_ method: String) -> Bool {
-        handlers[method] != nil
-    }
-
-    func getHandler(for method: String) -> RpcHandler? {
-        handlers[method]
-    }
-}
